Connect flask app to database.
