+++ docker-compose.yml
@@ -3,10 +3,27 @@ version: "3.7"
 services:
   postgres:
     image: postgres:12
-  ports:
-    - 5432:5432
-  volumes:
-    - "/tmp/hanabi-data:/var/lib/postgresql/data"
+
+    environment:
+      POSTGRES_PASSWORD: password
+    ports:
+      - 5432:5432
+    volumes:
+      - "/tmp/hanabi-data:/var/lib/postgresql/data"
+
+  migrations:
+    build:
+      context: .
+      target: devserver
+    image: hanabi-dev
+
+    entrypoint:
+      - flask
+      - db
+      - upgrade
+    environment:
+      DB_HOST: postgres
+      DB_PASSWORD: password
 
   server:
     build:
@@ -14,8 +31,11 @@ services:
       target: devserver
     image: hanabi-dev
 
+    environment:
+      DB_HOST: postgres
+      DB_PASSWORD: password
     ports:
       # Flask defaults to running on port 5000
       - 5000:5000
     volumes:
-    - ./:/opt/hanabi
+      - ./:/opt/hanabi
