+++ hanabi/static/hanabi_deck/Blue_2.png
