+++ hanabi/static/hanabi_deck/Blue_3.png
