+++ hanabi/static/hanabi_deck/Blue_4.png
