+++ hanabi/static/hanabi_deck/Blue_5.png
