+++ hanabi/static/hanabi_deck/Yellow_2.png
