+++ hanabi/static/scripts/game-in-session.js
