+++ hanabi/static/styles/game-in-session.css
