+++ hanabi/templates/game-in-session.html
