+++ requirements.txt
@@ -1,7 +1,11 @@
-Flask-WTF==0.14.3
 black==19.10b0
 flake8==3.7.9
 flask==1.1.1
+flask-migrate==2.5.3
+flask-sqlalchemy==2.4.1
+flask-wtf==0.14.3
 pre-commit==2.2.0
+psycopg2-binary==2.8.4
 pytest==5.4.1
+sqlalchemy==1.3.15
 wtforms==2.2.1
